+++ src/Position.cpp
@@ -32,8 +32,8 @@ void Position::setY(float y) {
 	this->y = y;
 }
 
-int Position::dist(Position p){
-	printf("calcul de sqrt((pow%d-%d,2)+(pow(%d-%d,2))\n", this->x, p.getX(), this->y, p.getY());
+float Position::dist(Position p){
+	printf("calcul de sqrt((pow%f-%f,2)+(pow(%f-%f,2))\n", this->x, p.getX(), this->y, p.getY());
 	return sqrt(
 		(pow(this->x-p.getX(), 2))+
 		(pow(this->y-p.getY(), 2))
